Revert "added google tag manager with cookiebot; specific event tagging done to work with Docusaurus as a SPA"
