+++ docusaurus/docusaurus.config.ts
@@ -33,18 +33,6 @@ const config: Config = {
         type: 'text/javascript',
       },
     },
-    // Google Tag Manager script (mirrors partnercenter-docs approach)
-    {
-      tagName: 'script',
-      attributes: {
-        type: 'text/javascript',
-      },
-      innerHTML: `(function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
-new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
-j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
-'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
-})(window,document,'script','dataLayer','${process.env.GTM_CONTAINER_ID || 'GTM-XXXXXXX'}');`,
-    },
   ],
 
   // Enable faster builds with Rspack bundler and persistent cache
@@ -67,6 +55,12 @@ j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
     [
       'classic',
       {
+        // ← GA4 configuration
+        gtag: {
+          trackingID: 'G-1Y49QBYD4L',  // your GA4 Measurement ID
+          anonymizeIP: true,           // optional: set to false to disable
+        },
+  
         docs: {
           sidebarPath: './sidebars.ts',
           // Serve docs at site root so "/" shows the docs with sidebar
@@ -143,19 +137,8 @@ j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
           title: 'Legal',
           items: [
             {
-              label: 'Privacy Policy',
-              to: '/legal/privacy-policy',
-            },
-            {
-              label: 'Terms of Use',
-              to: '/legal/terms',
-            },
-            {
               label: 'Cookie Declaration',
-              to: '/legal/privacy-policy#cookie-declaration',
-            },
-            {
-              html: '<a href="#cookie-settings" data-cookie-settings-link>Cookie Settings</a>',
+              to: '/cookie-declaration',
             },
           ],
         },
