--- docusaurus/src/pages/legal/privacy-policy.mdx
@@ -1,31 +1,0 @@
----
-title: Privacy Policy
-description: Privacy Policy for the Business App documentation site
----
-
-# Privacy Policy
-
-This documentation site ("Site") provides information about Business App. This Privacy Policy explains how this Site uses cookies and similar technologies and the limited data processed for analytics and performance.
-
-## Cookies and Consent
-
-This Site uses Cookiebot to manage cookie consent and to control when tags fire. You can review and update your consent at any time via the Cookie Settings link in the footer.
-
-<details id="cookie-declaration">
-<summary>View Cookie Declaration</summary>
-
-import CookieDeclaration from '../../components/CookieDeclaration';
-
-<CookieDeclaration />
-
-</details>
-
-## Analytics
-
-We use Google Tag Manager to deploy analytics tags, including Google Analytics 4 (GA4). Data is used in aggregate to improve Site content and navigation. IP anonymization may be enabled, and personal data is not used for profiling from this Site.
-
-## Contact
-
-If you have questions about this Privacy Policy, please contact the Business App team through your usual support channel.
-
-
