--- docusaurus/src/pages/legal/terms.mdx
@@ -1,14 +1,0 @@
----
-title: Terms of Use
-description: Terms of Use for the Business App documentation site
----
-
-# Terms of Use
-
-These Terms of Use govern your access to and use of this documentation site ("Site"). The Site provides documentation and learning resources about Business App. By using the Site, you agree to these terms.
-
-The Site is provided "as is" without warranties of any kind. Content is provided for informational purposes and may change without notice.
-
-If you have questions about these Terms, please contact the Business App team through your usual support channel.
-
-
